Runnable version! * Switched back from double to float for vector. seems to have better precision * Left-button is D not B... typo * VectorTest: properly tests normalize() * VectorTest: add testMult. testAdd sill TODO * Vector: fixed broken normalize()-method. * GraphicalUnit: representation of unit is now a Box * Unit: using float instead of double * Unit: small changes to steerClockwise/AntiClockwise

+++ project/src/model/Battlefield.java
@@ -21,7 +21,7 @@ public class Battlefield {
      * @param xWidth 
      * @param yWidth 
      */
-    public Battlefield(double xWidth, double yWidth) {
+    public Battlefield(float xWidth, float yWidth) {
         // Checking for errors
         if(xWidth == 0 && yWidth != 0) {
             xWidth = yWidth;
